fix: non auth stuff
